minor fix: convert data type

+++ pcError.cc
@@ -84,7 +84,7 @@ namespace pc {
       factor = desr_err[1] / sqrt(curr_err[1]*curr_err[1]
                                  +curr_err[2]*curr_err[2]
                                  +curr_err[3]*curr_err[3]);
-      h_new = h_old * pow(factor, 2.0/(2.0*(1.0+1.0-exp_m)+nsd));
+      h_new = h_old * pow(factor, 2.0/(2.0*(1.0+1.0-exp_m)+(double)nsd));
       //set new size
       apf::setScalar(elm_size, elm, 0, h_new);
       apf::destroyMeshElement(me);
